refactor: refactor member file path
